automated_price_tracking/pyproject.toml
[tool.poetry]
name = "automated-price-tracking"
version = "0.1.0"
description = ""
authors = ["BexTuychiev <bex@ibexprogramming.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.40.2"
pandas = "^2.2.3"
pydantic = "^2.10.2"
pydantic-settings = "^2.6.1"
sqlalchemy = "2.0.35"
aiohttp = "^3.11.8"
python-dotenv = "^1.0.1"
firecrawl-py = "^1.6.1"
requests = "^2.32.3"
plotly = "^5.24.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
